+++ modules/__init__.py
@@ -3,6 +3,7 @@ import importlib
 
 menu_ptr = []
 run_ptr = []
+pelican_settings = []
 master_redirections_dict = {}
 
 def sort_menu_by_priority():
@@ -37,6 +38,8 @@ for module in os.listdir('modules'):
             menu_ptr.append(imported_module.get_menu())
         if hasattr(imported_module, "run_module") and hasattr(imported_module, "get_priority"):
             run_ptr.append({"run_module":imported_module.run_module, "name": module, "priority": imported_module.get_priority()})
+        if hasattr(imported_module, "send_to_pelican"):
+            pelican_settings.append({"name": imported_module.send_to_pelican()})
         if hasattr(imported_module, "get_redirections"):
             redirections_list = imported_module.get_redirections()
             # Check list for conflicts