+++ modules/techniques/techniques.py
@@ -68,7 +68,7 @@ def generate_domain_markdown(domain, techniques_no_sub, tactics, side_nav_data):
 
         # Get technique table data and number of techniques
         data['technique_table'] = util.buildhelpers.get_technique_table_data(None, techhnique_list_no_sub_no_deprecated)
-        data['technique_list_len'] = str(len(techniques_no_sub[domain]))
+        data['technique_list_len'] = str(len(techhnique_list_no_sub_no_deprecated))
         data['subtechniques_len'] = util.buildhelpers.get_subtechnique_count(techhnique_list_no_sub_no_deprecated)
 
         # Get tactic-techniques table
@@ -80,8 +80,7 @@ def generate_domain_markdown(domain, techniques_no_sub, tactics, side_nav_data):
         with open(os.path.join(techniques_config.techniques_markdown_path, data['domain'] + "-techniques.md"), "w", encoding='utf8') as md_file:
             md_file.write(subs)
 
-        # Create the markdown for the enterprise groups in the STIX
-
+        # Create the markdown for techniques in the STIX
         for technique in techniques_no_sub[domain]:
             if 'revoked' not in technique or technique['revoked'] is False:
                 generate_technique_md(technique, domain, side_nav_data, tactics[domain])
@@ -184,10 +183,9 @@ def generate_data_for_md(technique_dict, technique, tactic_list, is_sub_techniqu
                     technique_dict['mtcs'].append(mtcs_dict)
 
         # Get initial reference list
-        reference_list = []
-        # Decleared as an object to be able to pass by reference
-        next_reference_number = {}
-        next_reference_number['value'] = 1
+        reference_list = {'current_number': 0}
+        
+        # Get initial reference list from technique object
         reference_list = util.buildhelpers.update_reference_list(reference_list, technique)
 
         dates = util.buildhelpers.get_created_and_modified_dates(technique)
@@ -206,20 +204,16 @@ def generate_data_for_md(technique_dict, technique, tactic_list, is_sub_techniqu
         # Get technique description with citations
         if technique.get("description") and not technique_dict['deprecated']:
 
-            citations_from_descr = util.buildhelpers.get_citations_from_descr(technique['description'])
-
-            technique_dict['descr'] = util.buildhelpers.replace_html_chars(markdown.markdown(technique['description']))
-            technique_dict['descr'] = util.buildhelpers.filter_urls(technique_dict['descr'])
-            technique_dict['descr'] = util.buildhelpers.get_descr_reference_sect(citations_from_descr, reference_list, next_reference_number, technique_dict['descr'])
+            technique_dict['descr'] = technique['description']
         
             # Get mitigation table
-            technique_dict['mitigation_table'] = get_mitigations_table_data(technique, reference_list, next_reference_number)
+            technique_dict['mitigation_table'] = get_mitigations_table_data(technique, reference_list)
             
             # Get related techniques
             technique_dict['rel_techniques_table'] = get_related_techniques_data(technique, tactic_list)
 
             # Get examples
-            technique_dict['examples_table'] = get_examples_table_data(technique, reference_list, next_reference_number)
+            technique_dict['examples_table'] = get_examples_table_data(technique, reference_list)
 
             # Get technique version
             if technique.get("x_mitre_version"):
@@ -293,7 +287,7 @@ def generate_data_for_md(technique_dict, technique, tactic_list, is_sub_techniqu
 
             # Get detection data
             if technique.get('x_mitre_detection'):
-                technique_dict['detection'] = get_detection_string(technique['x_mitre_detection'], reference_list, next_reference_number)
+                technique_dict['detection'] = technique['x_mitre_detection']
 
             # Get if technique is detectable by common defenses
             if technique.get('x_mitre_detectable_by_common_defenses'):
@@ -311,14 +305,11 @@ def generate_data_for_md(technique_dict, technique, tactic_list, is_sub_techniqu
             if technique.get('x_mitre_difficulty_for_adversary_explanation'):
                 technique_dict['diff_for_adv_exp'] = util.buildhelpers.replace_html_chars(technique['x_mitre_difficulty_for_adversary_explanation'])            
             
-            # Add reference for bottom part of technique page
-            if reference_list:
-                technique_dict['bottom_ref'] = util.buildhelpers.sort_reference_list(reference_list)
+            technique_dict['citations'] = reference_list
         
         else:
             if technique_dict['deprecated']:
-                technique_dict['descr'] = util.buildhelpers.replace_html_chars(markdown.markdown(technique['description'].split("\n")[0]))
-                technique_dict['descr'] = util.buildhelpers.filter_urls(technique_dict['descr'])
+                technique_dict['descr'] = technique['description']
 
         return technique_dict
 
@@ -351,7 +342,7 @@ def get_related_techniques_data(technique, tactic_list):
         technique_data = sorted(technique_data, key=lambda k: k['tactic_name'].lower())
     return technique_data
 
-def get_mitigations_table_data(technique, reference_list, next_reference_number):
+def get_mitigations_table_data(technique, reference_list):
     """Given a technique a reference list, find mitigations that mitigate
        technique and return list with mitigation data. Also modifies the 
        reference list if it finds a reference that is not on the list
@@ -377,7 +368,8 @@ def get_mitigations_table_data(technique, reference_list, next_reference_number)
                     row['name'] = mitigation['object']['name']
                     if mitigation['relationship'].get('description'):
                         # Get filtered description
-                        row['descr'] = util.buildhelpers.get_filtered_description(reference_list, next_reference_number, mitigation)
+                        reference_list = util.buildhelpers.update_reference_list(reference_list, mitigation['relationship'])
+                        row['descr'] = mitigation['relationship']['description']
              
                     mitigation_data.append(row)
     
@@ -385,7 +377,7 @@ def get_mitigations_table_data(technique, reference_list, next_reference_number)
         mitigation_data = sorted(mitigation_data, key=lambda k: k['name'].lower())
     return mitigation_data
     
-def get_examples_table_data(technique, reference_list, next_reference_number):
+def get_examples_table_data(technique, reference_list):
     """Given a technique object, find examples in malware using technique,
        tools using technique and groups using technique. Return list with
        example data
@@ -422,7 +414,8 @@ def get_examples_table_data(technique, reference_list, next_reference_number):
 
                     if example['relationship'].get('description'):
                         # Get filtered description
-                        row['descr'] = util.buildhelpers.get_filtered_description(reference_list, next_reference_number, example)
+                        reference_list = util.buildhelpers.update_reference_list(reference_list, example['relationship'])
+                        row['descr'] = example['relationship']['description']
 
                     example_data.append(row)
         
@@ -512,15 +505,11 @@ def get_techniques_list(techniques):
                 technique_dict['id'] = attack_id
                 technique_dict['stix_id'] = technique['id']
                 technique_dict['name'] = technique['name']
-                if technique['name'] == "DNSCalc":
-                    print(technique)
-                    exit()
                 technique_dict['description'] = technique['description']
 
                 if technique.get('kill_chain_phases'):
                     for elem in technique['kill_chain_phases']:
                         # Fill dict
-
                         if elem['phase_name'] not in technique_list:
                             technique_list[elem['phase_name']] = []
                             
@@ -530,21 +519,6 @@ def get_techniques_list(techniques):
         technique_list[key] = sorted(technique_list[key], key=lambda k: k['name'].lower())
     
     return technique_list
-
-
-def get_detection_string(detection, reference_list, next_reference_number):
-    """Given the technique's detection string, replace the citations and
-       add them to the reference_list is they are not found. Return
-       modified detection string
-    """
-
-    citations_from_descr = util.buildhelpers.get_citations_from_descr(detection)
-
-    filtered_detection = util.buildhelpers.replace_html_chars(markdown.markdown(detection))
-    filtered_detection = util.buildhelpers.filter_urls(filtered_detection)
-    filtered_detection = util.buildhelpers.get_descr_reference_sect(citations_from_descr, reference_list, next_reference_number, filtered_detection)
-    
-    return filtered_detection
 
 def get_subtechniques(technique):
     """Given a technique, return the ID and name of the subtechnique"""