+++ modules/website_build/__init__.py
@@ -1,5 +1,6 @@
 from . import website_build 
 from . import website_build_config
+from modules import site_config
 
 def get_priority():
     return website_build_config.priority